requirements.txt
alabaster==0.7.12
alembic==1.4.2
asgiref==3.2.3
attrs==19.3.0
Babel==2.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==5.0.1
Django==3.0.1
docker==4.2.0
docutils==0.15.2
entrypoints==0.3
flake8==3.7.9
Flask==1.1.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
idna==2.8
image==1.5.27
imagesize==1.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
numpy==1.17.4
packaging==19.2
pika==1.1.0
Pillow==6.2.1
pluggy==0.13.1
protobuf==3.11.2
psycopg2-binary==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
Pygments==2.5.2
pyparsing==2.4.5
pytest==5.3.2
pytest-cov==2.8.1
pytest-env==0.6.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
requests==2.22.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.15
sqlparse==0.3.0
urllib3==1.25.7
wcwidth==0.1.7
websocket-client==0.57.0
Werkzeug==1.0.0
